+++ tests/scripts/envs/pymssql.sh
@@ -1,0 +1,7 @@
+export PYTEST_MARKER="-m pymssql"
+export DOCKER_DEPS="mssql"
+export MSSQL_HOST="mssql"
+export MSSQL_USER="SA"
+export MSSQL_PASSWORD="Very(!)Secure"
+export WAIT_FOR_HOST="mssql"
+export WAIT_FOR_PORT=1433