+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+[Check the diff](https://github.com/elastic/apm-agent-python/compare/v2.2.1...master)
+
+ - added instrumentation for pymssql (#241)
+
 ## v2.2.1
 
 [Check the diff](https://github.com/elastic/apm-agent-python/compare/v2.2.0...v2.2.1)